+++ tests/integration/buildcmd_nodejs/test_build_esbuild.py
@@ -1,7 +1,8 @@
 import logging
 import os
+import shutil
 from pathlib import Path
-from unittest import skipIf
+from unittest import skipIf, skip
 
 import pytest
 from parameterized import parameterized, parameterized_class
@@ -177,3 +178,65 @@ class TestBuildCommand_EsbuildFunctionProperties(BuildIntegEsbuildBase):
             "architecture": architecture,
         }
         self._test_with_various_properties(overrides)
+
+
+@skip("Building in source option is not exposed yet. Stop skipping once it is.")
+class TestBuildCommand_BuildInSource_Esbuild(BuildIntegEsbuildBase):
+    is_nested_parent = False
+
+    def setUp(self):
+        super().setUp()
+        self.source_directories = []
+
+    def tearDown(self):
+        super().tearDown()
+        # clean up dependencies installed in source directories
+        for source in self.source_directories:
+            shutil.rmtree(os.path.join(source, "node_modules"), ignore_errors=True)
+
+    @parameterized.expand(
+        [
+            (True, True),  # build in source
+            (False, False),  # don't build in source
+            (None, False),  # use default for workflow (don't build in source)
+        ]
+    )
+    @pytest.mark.flaky(reruns=3)
+    def test_builds_successfully_without_local_dependencies(self, build_in_source, dependencies_expected_in_source):
+        self.template_path = os.path.join(self.test_data_path, "template_with_metadata.yaml")
+        codeuri = os.path.join(self.test_data_path, "Esbuild", "Node")
+        self.source_directories = [codeuri]
+
+        self._test_with_default_package_json(
+            build_in_source=build_in_source,
+            runtime="nodejs16.x",
+            code_uri=codeuri,
+            handler="main.lambdaHandler",
+            architecture="x86_64",
+            use_container=False,
+            expected_files={"main.js", "main.js.map"},
+        )
+
+        # check whether dependencies were installed in source dir
+        self.assertEqual(os.path.isdir(os.path.join(codeuri, "node_modules")), dependencies_expected_in_source)
+
+    @pytest.mark.flaky(reruns=3)
+    def test_builds_successfully_with_local_dependency(self):
+        self.template_path = os.path.join(self.test_data_path, "template_with_metadata.yaml")
+        codeuri = os.path.join(self.test_data_path, "Esbuild", "NodeWithLocalDependency")
+        self.source_directories = [codeuri]
+        runtime = "nodejs16.x"
+        architecture = "x86_64"
+
+        self._test_with_default_package_json(
+            build_in_source=True,
+            runtime=runtime,
+            code_uri=codeuri,
+            handler="main.lambdaHandler",
+            architecture=architecture,
+            use_container=False,
+            expected_files={"main.js", "main.js.map"},
+        )
+
+        # check whether dependencies were installed in source dir
+        self.assertEqual(os.path.isdir(os.path.join(codeuri, "node_modules")), True)
